% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bootstrap.cc}
\alias{bootstrap.cc}
\title{A helper function to bootstrap risk function correcting for misclassification between censoring and competing events}
\usage{
bootstrap.cc(
  B,
  bdat1,
  t = "t",
  delta = "delta",
  tau,
  p = "p",
  sep = "sep",
  seed = 456
)
}
\arguments{
\item{B}{number of bootstrap samples to draw}

\item{bdat1}{input data source containing main study data}

\item{t}{character string denoting name of time variable}

\item{delta}{character string denoting name of event type variable}

\item{tau}{final follow-up time for which risks and standard errors will be reported}

\item{p}{a character string denoting name of the variable containing the probability that person i is in care elsewhere}

\item{sep}{a character string denoting the name of the variable containing the standard error of the estimated probability that person i is in care elsewhere}

\item{seed}{seed}

\item{weights}{character string denoting name of variable containing weights (e.g., survey sampling weights, inverse probability weights, etc)}
}
\value{
standard error of the cumulative incidence (i.e., risk) function at time tau
}
\description{
A helper function to bootstrap risk function correcting for misclassification between censoring and competing events
}
\examples{
 se <- bootstrap.cc(B, data, t = "t", delta = "j", tau, p = "p", sep = "selogitp", seed=123)
 
}
