% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{comp.cens.cor}
\alias{comp.cens.cor}
\title{A function to estimate a nonparametric cumulative incidence function accounting for misclassification between competing events and censoring events
#'}
\usage{
comp.cens.cor(
  data = data,
  tau,
  t = "t",
  delta = "delta",
  weights = "NULL",
  p = "p"
)
}
\arguments{
\item{data}{input data source containing main study data, specifically time variable and outcome type}

\item{tau}{final follow-up time}

\item{t}{character string denoting name of time variable}

\item{delta}{character string denoting name of event type variable}

\item{weights}{character string denoting name of variable containing weights (e.g., survey sampling weights, inverse probability weights, etc)}

\item{p}{a character string denoting name of the variable containing the probability that person i is in care elsewhere}
}
\value{
data frame with event times and cumulative incidence function for outcome of interest
}
\description{
A function to estimate a nonparametric cumulative incidence function accounting for misclassification between competing events and censoring events
#'
}
\examples{
 prop<-comp.cens.cor(data=art, tau=2*365.25, t="t", delta="j", p=0.24, N=1000)
}
