% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{risk}
\alias{risk}
\title{A helper function to estimate a nonparametric cumulative incidence function to the level needed for plotting}
\usage{
risk(
  data = data,
  t = "t",
  delta = "y",
  level = 1,
  tau = 4,
  weights = "NULL",
  subvar = "NULL",
  subval = NA
)
}
\arguments{
\item{data}{input data source containing main study data, specifically time variable and outcome type}

\item{t}{character string denoting name of time variable}

\item{delta}{character string denoting name of event type variable}

\item{tau}{final follow-up time}

\item{weights}{character string denoting name of variable containing weights (e.g., survey sampling weights, inverse probability weights, etc)}

\item{subvar}{a character string denoting the name of any subsetting variable (not called by other functions in this package)}

\item{subval}{the level of subvar of interest}
}
\value{
data frame with event times and cumulative incidence function for outcome of interest
}
\description{
A helper function to estimate a nonparametric cumulative incidence function to the level needed for plotting
}
\examples{
 r1<-risk(data=part, t="t", delta="y", tau=tau, weights="new_weight")
}
